use licheng delay

+++ USER/clock.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
 void systick_config(void);
-void delay_us(unsigned long us);
+void delay_us(uint32_t us);
